requirements.txt
pandas
scikit-learn
imblearn
joblib
numpy
scipy
seaborn
fastapi
requests
pyyaml